backend/Elfezya_Bgd/requirements.txt
Django==5.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-storages==1.14.6
django-allauth==65.13.1
django-nested-admin==4.1.6
python-dotenv==1.2.1

boto3==1.40.55
Pillow==11.2.1
google-auth==2.41.1
requests==2.32.5

gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.2.0

PyMySQL==1.1.2
psycopg[binary]==3.2.9
